+++ lib/model/spmmax_pooling/src/spmmax_pooling_kernel.h
@@ -7,10 +7,10 @@ extern "C" {
 #endif
 
 int spmmax_pooling_forward_kernel(int batch_size, int num_grids, int feature_size, int num_rois, float* x_data,
-                                  float* shapes_data, float* rois_data, float* output_data, int64_t* max_ids_data);
+                                  float* shapes_data, float* rois_data, float* output_data, int* max_ids_data);
 
 int spmmax_pooling_backward_kernel(int batch_size, int num_grids, int feature_size, int num_rois, float* grad_input_data,
-                                   float* grad_output_data, int64_t* max_ids_data);
+                                   float* grad_output_data, int* max_ids_data);
 
 #ifdef __cplusplus
 }
